move updateShortcutsActor to Dispatcher
