+++ ui/shortcuts/src/test/java/org/cru/godtools/shortcuts/GodToolsShortcutManagerDispatcherTest.kt
@@ -1,22 +1,28 @@
 package org.cru.godtools.shortcuts
 
+import android.os.Build
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import java.util.Locale
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.test.TestCoroutineScope
+import kotlinx.coroutines.test.runBlockingTest
 import org.cru.godtools.base.Settings
 import org.junit.After
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
+import org.junit.runner.RunWith
 import org.mockito.kotlin.clearInvocations
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verifyBlocking
 import org.mockito.kotlin.verifyNoInteractions
 import org.mockito.kotlin.verifyNoMoreInteractions
+import org.robolectric.annotation.Config
 
+@RunWith(AndroidJUnit4::class)
 @OptIn(ExperimentalCoroutinesApi::class)
 class GodToolsShortcutManagerDispatcherTest {
     private lateinit var settings: Settings
@@ -47,6 +53,8 @@ class GodToolsShortcutManagerDispatcherTest {
 
     @Test
     fun verifyUpdatePendingToolShortcuts() {
+        dispatcher.updateShortcutsActor.close()
+
         // update doesn't trigger before requested
         coroutineScope.advanceUntilIdle()
         coroutineScope.advanceTimeBy(DELAY_UPDATE_PENDING_SHORTCUTS)
@@ -71,4 +79,79 @@ class GodToolsShortcutManagerDispatcherTest {
         coroutineScope.advanceUntilIdle()
         verifyNoMoreInteractions(shortcutManager)
     }
+
+    // region updateShortcutsActor
+    @Test
+    @Config(sdk = [Build.VERSION_CODES.N_MR1, Config.NEWEST_SDK])
+    fun verifyUpdateExistingShortcutsOnPrimaryLanguageUpdate() {
+        dispatcher.updatePendingShortcutsActor.close()
+        assertUpdateExistingShortcutsInitialUpdate()
+
+        // trigger a primary language update
+        assertTrue(primaryLanguageFlow.tryEmit(Locale.ENGLISH))
+        verifyNoInteractions(shortcutManager)
+        coroutineScope.advanceUntilIdle()
+        verifyBlocking(shortcutManager) { updateShortcuts() }
+        verifyNoMoreInteractions(shortcutManager)
+    }
+
+    @Test
+    @Config(sdk = [Build.VERSION_CODES.N_MR1, Config.NEWEST_SDK])
+    fun verifyUpdateExistingShortcutsOnParallelLanguageUpdate() {
+        dispatcher.updatePendingShortcutsActor.close()
+        assertUpdateExistingShortcutsInitialUpdate()
+
+        // trigger a primary language update
+        assertTrue(parallelLanguageFlow.tryEmit(null))
+        verifyNoInteractions(shortcutManager)
+        coroutineScope.advanceUntilIdle()
+        verifyBlocking(shortcutManager) { updateShortcuts() }
+        verifyNoMoreInteractions(shortcutManager)
+    }
+
+    @Test
+    @Config(sdk = [Build.VERSION_CODES.N_MR1, Config.NEWEST_SDK])
+    fun verifyUpdateExistingShortcutsAggregateMultiple() = runBlockingTest {
+        dispatcher.updatePendingShortcutsActor.close()
+        assertUpdateExistingShortcutsInitialUpdate()
+
+        // trigger multiple updates simultaneously, it should aggregate to a single update
+        assertTrue(dispatcher.updateShortcutsActor.trySend(Unit).isSuccess)
+        assertTrue(primaryLanguageFlow.tryEmit(Locale.ENGLISH))
+        assertTrue(parallelLanguageFlow.tryEmit(null))
+        coroutineScope.advanceTimeBy(DELAY_UPDATE_SHORTCUTS - 4000)
+        verifyNoInteractions(shortcutManager)
+        assertTrue(dispatcher.updateShortcutsActor.trySend(Unit).isSuccess)
+        assertTrue(dispatcher.updateShortcutsActor.trySend(Unit).isSuccess)
+        assertTrue(primaryLanguageFlow.tryEmit(Locale.ENGLISH))
+        assertTrue(parallelLanguageFlow.tryEmit(null))
+        coroutineScope.advanceTimeBy(DELAY_UPDATE_SHORTCUTS - 1)
+        verifyNoInteractions(shortcutManager)
+        coroutineScope.advanceUntilIdle()
+        verifyBlocking(shortcutManager) { updateShortcuts() }
+    }
+
+    @Test
+    @Config(sdk = [Config.OLDEST_SDK, Build.VERSION_CODES.N])
+    fun verifyUpdateExistingShortcutsNotAvailableForOldSdks() {
+        coroutineScope.advanceUntilIdle()
+        assertTrue(
+            "Ensure actor can still accept requests, even though they are no-ops",
+            dispatcher.updateShortcutsActor.trySend(Unit).isSuccess
+        )
+        coroutineScope.advanceUntilIdle()
+        verifyNoInteractions(shortcutManager)
+    }
+
+    private fun assertUpdateExistingShortcutsInitialUpdate() {
+        // ensure update shortcuts is initially delayed
+        coroutineScope.advanceTimeBy(DELAY_UPDATE_SHORTCUTS - 1)
+        verifyNoInteractions(shortcutManager)
+        coroutineScope.advanceTimeBy(1)
+        verifyBlocking(shortcutManager) { updateShortcuts() }
+        coroutineScope.advanceUntilIdle()
+        verifyNoMoreInteractions(shortcutManager)
+        clearInvocations(shortcutManager)
+    }
+    // endregion updateShortcutsActor
 }
